re PR middle-end/90518 (ICE: in emit_move_insn, at expr.c:3745 in gcc.dg/gimplefe-40.c)
2019-05-20  Richard Biener  <rguenther@suse.de>

	PR testsuite/90518
	* gcc.dg/gimplefe-40.c: Restrict to targets with appropriate
	vector support.
	* gcc.dg/gimplefe-41.c: Likewise.

From-SVN: r271396

+++ gcc/testsuite/gcc.dg/gimplefe-41.c
@@ -1,8 +1,8 @@
-/* { dg-do compile } */
+/* { dg-do compile { target { vect_double && vect_long_long } } } */
 /* { dg-options "-fgimple -Wno-psabi -w" } */
 
 typedef double __v2df __attribute__ ((__vector_size__ (16)));
-typedef unsigned long __v2di __attribute__ ((__vector_size__ (16)));
+typedef unsigned long long __v2di __attribute__ ((__vector_size__ (16)));
 
 __v2df __GIMPLE (ssa)
 _mm_add_sd (__v2df x, __v2df y)
